# src/server/api/go/docs/swagger.yaml
basePath: /api/v1
definitions:
  handler.ConnectToSpaceReq:
    properties:
      space_id:
        example: 123e4567-e89b-12d3-a456-426614174000
        format: uuid
        type: string
    required:
    - space_id
    type: object
...
  handler.CreateSessionReq:
    properties:
      configs:
        additionalProperties: true
        type: object
      space_id:
        example: 123e4567-e89b-12d3-a456-42661417
        format: uuid
        type: string
    type: object
...
  handler.SendMessageReq:
    properties:
      format:
        enum:
        - acontext
        - openai
        - langchain
        - anthropic
        example: openai
        type: string
      parts:
        items:
          $ref: '#/definitions/service.PartIn'
        type: array
      role:
        enum:
        - user
        - assistant
        - system
        - tool
        - function
        example: user
        type: string
    required:
    - parts
    - role
    type: object
...
  handler.UpdateSessionConfigsReq:
    properties:
      configs:
        additionalProperties: true
        type: object
    type: object
...
  model.Message:
    properties:
      children:
        items:
          $ref: '#/definitions/model.Message'
        type: array
      created_at:
        type: string
      id:
        type: string
      parent:
        $ref: '#/definitions/model.Message'
      parent_id:
        type: string
      parts:
        items:
          type: object
        type: array
      role:
        type: string
      session:
        allOf:
        - $ref: '#/definitions/model.Session'
        description: Message <-> Session
      session_id:
        type: string
      session_task_process_status:
        type: string
      task:
        allOf:
        - $ref: '#/definitions/model.Task'
        description: Message <-> Task
      task_id:
        type: string
      updated_at:
        type: string
    type: object
  model.Project:
    properties:
      configs:
        type: object
      created_at:
        type: string
      id:
        type: string
      sessions:
        description: Project <-> Session
        items:
          $ref: '#/definitions/model.Session'
        type: array
      spaces:
        description: Project <-> Space
        items:
          $ref: '#/definitions/model.Space'
        type: array
      updated_at:
        type: string
    type: object
  model.Session:
    properties:
      configs:
        type: object
      created_at:
        type: string
      id:
        type: string
      messages:
        description: Session <-> Message
        items:
          $ref: '#/definitions/model.Message'
        type: array
      project:
        allOf:
        - $ref: '#/definitions/model.Project'
        description: Session <-> Project
      project_id:
        type: string
      space:
        allOf:
        - $ref: '#/definitions/model.Space'
        description: Session <-> Space
      space_id:
        type: string
      tasks:
        description: Session <-> Task
        items:
          $ref: '#/definitions/model.Task'
        type: array
      updated_at:
        type: string
    type: object
  model.Space:
    properties:
      configs:
        type: object
      created_at:
        type: string
      id:
        type: string
      project:
        allOf:
        - $ref: '#/definitions/model.Project'
        description: Space <-> Project
      project_id:
        type: string
      sessions:
        description: Space <-> Session
        items:
          $ref: '#/definitions/model.Session'
        type: array
      updated_at:
        type: string
    type: object
  model.Task:
    properties:
      created_at:
        type: string
      id:
        type: string
      is_planning_task:
        type: boolean
      messages:
        description: Task <-> Message (one-to-many)
        items:
          $ref: '#/definitions/model.Message'
        type: array
      session:
        allOf:
        - $ref: '#/definitions/model.Session'
        description: Task <-> Session
      session_id:
        type: string
      space_digested:
        type: boolean
      task_data:
        type: object
      task_order:
        type: integer
      task_status:
        type: string
      updated_at:
        type: string
    type: object
  serializer.Response:
    properties:
      code:
        type: integer
      data: {}
      error:
        type: string
      msg:
        type: string
    type: object
  service.GetMessagesOutput:
    properties:
      has_more:
        type: boolean
      items:
        items:
          $ref: '#/definitions/model.Message'
        type: array
      next_cursor:
        type: string
      public_urls:
        additionalProperties:
          $ref: '#/definitions/service.PublicURL'
        description: file_name -> url
        type: object
    type: object
  service.PartIn:
    properties:
      file_field:
        description: File field name in the form
        type: string
      meta:
        additionalProperties: true
        description: '[Optional] metadata'
        type: object
      text:
        description: Text sharding
        type: string
      type:
        description: '"text" | "image" | ...'
        enum:
        - text
        - image
        - audio
        - video
        - file
        - tool-call
        - tool-result
        - data
        type: string
    required:
    - type
    type: object
  service.PublicURL:
    properties:
      expire_at:
        type: string
      url:
        type: string
    type: object
info:
  contact: {}
  description: API for Acontext.
  title: Acontext API
  version: "1.0"
paths:
...
  /session:
    get:
      consumes:
      - application/json
      description: Get all sessions under a project, optionally filtered by space_id
      parameters:
      - description: Space ID to filter sessions
        format: uuid
        in: query
        name: space_id
        type: string
      - description: Filter sessions not connected to any space (true/false)
        example: "true"
        in: query
        name: not_connected
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/serializer.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Session'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get sessions
      tags:
      - session
    post:
      consumes:
      - application/json
      description: Create a new session under a space
      parameters:
      - description: CreateSession payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.CreateSessionReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/serializer.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Session'
              type: object
      security:
      - BearerAuth: []
      summary: Create session
      tags:
      - session
  /session/{session_id}:
    delete:
      consumes:
      - application/json
      description: Delete a session by id
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.Response'
      security:
      - BearerAuth: []
      summary: Delete session
      tags:
      - session
  /session/{session_id}/configs:
    get:
      consumes:
      - application/json
      description: Get session configs by id
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/serializer.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Session'
              type: object
      security:
      - BearerAuth: []
      summary: Get session configs
      tags:
      - session
    put:
      consumes:
      - application/json
      description: Update session configs by id
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      - description: UpdateSessionConfigs payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSessionConfigsReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.Response'
      security:
      - BearerAuth: []
      summary: Update session configs
      tags:
      - session
  /session/{session_id}/connect_to_space:
    post:
      consumes:
      - application/json
      description: Connect a session to a space by id
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      - description: ConnectToSpace payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.ConnectToSpaceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/serializer.Response'
      security:
      - BearerAuth: []
      summary: Connect session to space
      tags:
      - session
  /session/{session_id}/messages:
    get:
      consumes:
      - application/json
      description: Get messages from session. Default format is openai. Can convert
        to acontext (original), langchain, or anthropic format.
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      - description: Limit of messages to return, default 20. Max 200.
        in: query
        name: limit
        type: integer
      - description: Cursor for pagination. Use the cursor from the previous response
          to get the next page.
        in: query
        name: cursor
        type: string
      - description: Whether to return asset public url, default is true
        in: query
        name: with_asset_public_url
        type: string
      - description: 'Format to convert messages to: acontext (original), openai (default),
          langchain, anthropic.'
        enum:
        - acontext
        - openai
        - langchain
        - anthropic
        in: query
        name: format
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/serializer.Response'
            - properties:
                data:
                  $ref: '#/definitions/service.GetMessagesOutput'
              type: object
      security:
      - BearerAuth: []
      summary: Get messages from session
      tags:
      - session
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: 'Supports JSON and multipart/form-data. In multipart mode: the
        payload is a JSON string placed in a form field. The format parameter indicates
        the format of the input message (default: openai, same as GET).'
      parameters:
      - description: Session ID
        format: uuid
        in: path
        name: session_id
        required: true
        type: string
      - description: 'SendMessage payload (Content-Type: application/json)'
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.SendMessageReq'
      - description: 'SendMessage payload (Content-Type: multipart/form-data)'
        in: formData
        name: payload
        type: string
      - description: When uploading files, the field name must correspond to parts[*].file_field.
        in: formData
        name: file
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/serializer.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.Message'
              type: object
      security:
      - BearerAuth: []
      summary: Send message to session
      tags:
      - session
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Project Bearer token (e.g., "Bearer sk-ac-xxxx")
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"